// app/src/lib/types.ts
export type Settings = {
  id: string;
  userId: string;
  headline: string;
  experience: number;
  location: string;
  portfolio: string;
  github: string;
  linkedin: string;
  skills: string;
  achievements: string;
  customPrompt: string;
  gmailConnected: boolean;
  gmailTokens: string | null;
  createdAt: number;
  updatedAt: number;
};